move function under quoter
